+++ test/project/uiengine/pages/testcases/custom-fragment-and-template/page.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  template: 'template.pug',
+  fragment: 'fragment.pug',
+  tags: ['Template'],
+  context: {
+    myContextVariable: 'This is my context',
+    myFragmentVariable: 'This is my fragment'
+  }
+}