+++ test/project/uiengine/pages/testcases/custom-fragment/page.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  fragment: 'fragment.pug',
+  tags: ['Template'],
+  context: {
+    myFragmentVariable: 'This is my fragment'
+  }
+}